Reworked http.ts
